requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-recaptcha==4.1.0
django-two-factor-auth==1.17.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
loguru==0.7.3
packaging==25.0
phonenumbers==8.13.55
pillow==11.2.1
psycopg2-binary==2.9.10
pypng==0.20220715.0
python-dotenv==1.1.0
qrcode==7.4.2
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
loguru==0.7.3
